+++ projects/demo/src/modules/components/input-phone-international/examples/2/index.ts
@@ -4,10 +4,10 @@ import {FormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
 import {TuiTextfield} from '@taiga-ui/core';
-import {TuiInputPhoneInternational} from '@taiga-ui/experimental';
 import {type TuiCountryIsoCode} from '@taiga-ui/i18n';
 import {
     tuiInputPhoneInternationalOptionsProvider,
+    TuiInputPhoneInternational,
     TuiSortCountriesPipe,
 } from '@taiga-ui/kit';
 import {getCountries} from 'libphonenumber-js';
