+++ projects/demo/src/modules/components/input-phone-international/examples/3/index.ts
@@ -5,10 +5,10 @@ import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
 import {TuiDropdownMobile} from '@taiga-ui/addon-mobile';
 import {TuiTextfield} from '@taiga-ui/core';
-import {TuiInputPhoneInternational} from '@taiga-ui/experimental';
 import {type TuiCountryIsoCode} from '@taiga-ui/i18n';
 import {
     tuiInputPhoneInternationalOptionsProvider,
+    TuiInputPhoneInternational,
     TuiSortCountriesPipe,
 } from '@taiga-ui/kit';
 import {getCountries} from 'libphonenumber-js';
