+++ projects/demo/src/modules/components/input-phone-international/examples/4/index.ts
@@ -3,9 +3,12 @@ import {FormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
 import {TuiIcon, TuiTextfield} from '@taiga-ui/core';
-import {TuiInputPhoneInternational} from '@taiga-ui/experimental';
 import {type TuiCountryIsoCode} from '@taiga-ui/i18n';
-import {tuiInputPhoneInternationalOptionsProvider, TuiTooltip} from '@taiga-ui/kit';
+import {
+    tuiInputPhoneInternationalOptionsProvider,
+    TuiInputPhoneInternational,
+    TuiTooltip,
+} from '@taiga-ui/kit';
 import {defer} from 'rxjs';
 
 @Component({
