+++ projects/demo/src/modules/components/input-phone-international/examples/5/index.ts
@@ -3,9 +3,11 @@ import {FormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
 import {TuiTextfield} from '@taiga-ui/core';
-import {TuiInputPhoneInternational} from '@taiga-ui/experimental';
 import {type TuiCountryIsoCode} from '@taiga-ui/i18n';
-import {tuiInputPhoneInternationalOptionsProvider} from '@taiga-ui/kit';
+import {
+    tuiInputPhoneInternationalOptionsProvider,
+    TuiInputPhoneInternational,
+} from '@taiga-ui/kit';
 import {getCountries} from 'libphonenumber-js';
 import {defer} from 'rxjs';
