+++ projects/demo/src/modules/components/input-phone-international/examples/6/index.ts
@@ -8,9 +8,7 @@ import metadata from 'libphonenumber-js/max/metadata';
 
 @Component({
     imports: [MaskitoPipe],
-    template: `
-        Phone: {{ rawValue | maskito: mask }}
-    `,
+    template: 'Phone: {{ rawValue | maskito: mask }}',
     encapsulation,
     changeDetection,
     host: {
