+++ projects/demo/src/modules/components/input-phone-international/index.html
@@ -1,6 +1,6 @@
 <tui-doc-page
     header="InputPhoneInternational"
-    package="EXPERIMENTAL"
+    package="KIT"
     type="components"
 >
     <ng-template pageTab>
