+++ projects/demo/src/modules/components/input-phone-international/index.ts
@@ -7,9 +7,12 @@ import {changeDetection} from '@demo/emulate/change-detection';
 import {DemoRoute} from '@demo/routes';
 import {TuiDemo} from '@demo/utils';
 import {TuiDropdown, TuiIcon, TuiTextfield} from '@taiga-ui/core';
-import {TuiInputPhoneInternational} from '@taiga-ui/experimental';
 import {type TuiCountryIsoCode} from '@taiga-ui/i18n';
-import {tuiInputPhoneInternationalOptionsProvider, TuiTooltip} from '@taiga-ui/kit';
+import {
+    TuiInputPhoneInternational,
+    tuiInputPhoneInternationalOptionsProvider,
+    TuiTooltip,
+} from '@taiga-ui/kit';
 import {getCountries} from 'libphonenumber-js';
 
 @Component({
