+++ projects/experimental/components/index.ts
@@ -1,2 +1,1 @@
-export * from '@taiga-ui/experimental/components/input-phone-international';
 export * from '@taiga-ui/experimental/components/search-results';
