--- projects/experimental/components/input-phone-international/index.ts
@@ -1,1 +1,0 @@
-export * from './input-phone-international.component';
