--- projects/experimental/components/input-phone-international/input-phone-international.template.html
@@ -1,70 +1,0 @@
-<ng-container *tuiTextfieldContent>
-    <button
-        appearance="textfield"
-        tabindex="-1"
-        tuiButton
-        tuiChevron
-        type="button"
-        class="t-ipi-select"
-        [attr.data-mode]="mode()"
-        [class.t-ipi-select_readonly]="readOnly()"
-        [disabled]="disabled()"
-        [size]="size()"
-        [tuiAppearanceFocus]="open()"
-        (click.prevent)="interactive() && open.set(!open())"
-        (pointerdown.stop)="(0)"
-    >
-        <img
-            class="t-ipi-flag"
-            [alt]="names()?.[code()]"
-            [class.t-ipi-flag_small]="size() === 's'"
-            [src]="code() | tuiFlag"
-        />
-    </button>
-</ng-container>
-<ng-template #filter>
-    <div class="t-ipi-search">
-        <tui-textfield
-            [iconStart]="icons.search"
-            [tuiTextfieldSize]="size() === 's' ? 's' : 'm'"
-        >
-            <input
-                autocomplete="off"
-                tuiTextfield
-                type="text"
-                [ngModel]="search()"
-                [ngModelOptions]="{standalone: true}"
-                [placeholder]="label()"
-                [tuiAutoFocus]="!ios"
-                (keydown.arrowDown)="list.get(0)?.nativeElement?.focus()"
-                (ngModelChange)="search.set($event)"
-            />
-        </tui-textfield>
-    </div>
-</ng-template>
-<ng-container *tuiTextfieldDropdown>
-    <!-- To keep it under local injector -->
-    @if (countrySearch) {
-        <ng-container *ngTemplateOutlet="filter" />
-    }
-    <tui-data-list class="t-ipi-options">
-        @for (item of filtered(); track $index) {
-            <button
-                new
-                tuiOption
-                type="button"
-                [tuiCell]="size()"
-                (click)="onItemClick(item.iso)"
-            >
-                <img
-                    alt=""
-                    class="t-ipi-flag"
-                    [class.t-ipi-flag_small]="size() === 's'"
-                    [src]="item.iso | tuiFlag"
-                />
-                <span tuiTitle>{{ item.name }}</span>
-                <span class="t-ipi-code">{{ item.code }}</span>
-            </button>
-        }
-    </tui-data-list>
-</ng-container>
