--- projects/experimental/components/input-phone-international/ng-package.json
@@ -1,5 +1,0 @@
-{
-    "lib": {
-        "entryFile": "index.ts"
-    }
-}
