+++ projects/kit/components/input-phone-international/input-phone-international.style.less
@@ -1,37 +1,94 @@
 @import '@taiga-ui/core/styles/taiga-ui-local';
 
-.t-select {
-    inline-size: 5.625rem;
-    flex: none;
+[tuiInputPhoneInternational][tuiInputPhoneInternational] {
+    left: var(--t-offset);
+    border-top-left-radius: 0;
+    border-bottom-left-radius: 0;
+    inline-size: calc(100% - var(--t-offset));
 
-    &[data-size='m'] {
-        inline-size: 5rem;
-    }
-
-    &[data-size='s'] {
-        inline-size: 4rem;
+    & + label {
+        padding-inline-start: var(--t-offset);
     }
 }
 
-.t-flag {
+tui-textfield[data-size='s'] {
+    --t-offset: 4.125rem;
+
+    .t-ipi-flag {
+        margin: 0 0.1875rem;
+    }
+}
+
+tui-textfield[data-size='m'] {
+    --t-offset: 4.875rem;
+
+    .t-ipi-flag {
+        margin: 0 -0.1875rem;
+    }
+}
+
+tui-textfield[data-size='l'] {
+    --t-offset: 5.25rem;
+
+    .t-ipi-flag {
+        margin: 0 -0.1875rem;
+    }
+}
+
+tui-textfield .t-ipi-select {
+    position: absolute;
+    left: 0;
+    border-radius: inherit;
+    border-top-right-radius: 0;
+    border-bottom-right-radius: 0;
+
+    &_readonly {
+        pointer-events: none;
+    }
+}
+
+.t-ipi-flag {
     inline-size: 1.75rem;
     block-size: 1.75rem;
     border-radius: 100%;
+
+    &_small {
+        inline-size: 1.25rem;
+        block-size: 1.25rem;
+    }
 }
 
-.t-name {
-    margin: 0 auto 0 0.75rem;
-}
-
-.t-code {
+.t-ipi-code {
     color: var(--tui-text-secondary);
-    margin-inline-end: 0.25rem;
 }
 
-.t-search {
+.t-ipi-search {
     position: sticky;
-    top: 0.375rem;
+    top: 0;
+    z-index: 1;
     background: var(--tui-background-elevation-3);
-    box-shadow: 0 -1rem var(--tui-background-elevation-3);
-    margin: 0.375rem 0.375rem 0;
+    padding: 0.375rem 0.375rem 0;
+
+    input {
+        .tui-prevent-ios-scroll();
+    }
+}
+
+tui-dropdown-mobile,
+tui-sheet-dialog {
+    .t-ipi-search {
+        background: var(--tui-background-elevation-1);
+    }
+
+    .t-ipi-options:not(:first-child) {
+        min-block-size: calc(100 * var(--tui-viewport-vh) - 8.75rem);
+    }
+}
+
+tui-sheet-dialog {
+    .t-ipi-search {
+        top: 4.5rem;
+        padding-inline-start: 0;
+        padding-inline-end: 0;
+    }
 }
