+++ projects/layout/components/form/form.styles.less
@@ -41,8 +41,7 @@
     > fieldset {
         display: grid;
         gap: inherit;
-        grid-auto-flow: column;
-        grid-auto-columns: 1fr;
+        grid-template-columns: repeat(auto-fit, minmax(18rem, 1fr));
         border: none;
         padding: 0;
         margin: 0;
